new small modification to setup.py with additional mods for distribution and installation handling

+++ setup.py
@@ -22,9 +22,19 @@ import shutil
 from three2two import run3to2,run3to3
 
 # check whether a build directory is available
-if not os.path.isdir('lingpy_build'):
+if not os.path.isdir('lingpy_build') and 'install' in sys.argv:
     os.mkdir('lingpy_build')
     os.mkdir('lingpy_build/lingpy')
+
+# check for install and organize the manifest.in-file
+if 'install' in sys.argv:
+    m = open('manifest_build.in','r').read()
+else:
+    m = open('manifest_dist.in','r').read()
+f = open('MANIFEST.in','w')
+f.write(m)
+f.close()
+
 
 # check for specific features
 with_c = False
@@ -35,18 +45,25 @@ for i,arg in enumerate(sys.argv):
         break
 
 extra = {}
+
+# setup package name etc as a default
+pkgname = 'lingpy'
+pkg_dir = {'':'.'}
+pkg_location = '.'
+
 if sys.version_info >= (3,):
     extra['use_2to3'] = False
     this_version = "3"
-    pkg_location = 'lingpy_build'
-    pkgname = 'lingpy'
-    pkg_dir = {'':'lingpy_build'}
-    run3to3()
 
     # import lingpy from lingpy_build folder
-    sys.path = ['lingpy_build/'] + sys.path
-    from lingpy import *
-    rc(schema='asjp')
+    if 'install' in sys.argv:
+        pkg_location = 'lingpy_build'
+        pkgname = 'lingpy'
+        pkg_dir = {'':'lingpy_build'}
+        run3to3()
+        sys.path = ['lingpy_build/'] + sys.path
+        from lingpy import *
+        rc(schema='asjp')
     
 else:
     # make a specific directory for lingpy2
@@ -54,14 +71,16 @@ else:
 
     run3to2()
     pkgname = 'lingpy'
-    # replace manifest path
-    pkg_location = 'lingpy_build'
-    pkg_dir = {'':'lingpy_build'}
+
 
     # import lingpy2 to compile the data
-    sys.path = ['lingpy_build/'] + sys.path
-    from lingpy import *
-    rc(schema='asjp')
+    if 'install' in sys.argv:
+        # replace manifest path
+        pkg_location = 'lingpy_build'
+        pkg_dir = {'':'lingpy_build'}
+        sys.path = ['lingpy_build/'] + sys.path
+        from lingpy import *
+        rc(schema='asjp')
 
 # set up extension modules
 if 'install' in sys.argv or 'bdist_egg' in sys.argv:
@@ -102,7 +121,7 @@ else:
     extension_modules = []
 
 # make global name of this version
-thisversion = "2.1"
+thisversion = "2.2"
 setup(
         name = pkgname,
         version = thisversion,
@@ -131,7 +150,8 @@ setup(
 
 # remove the build directory in order to prevent that it leads to confusion
 # when installing lingpy in both the py2 and the py3 version
-print("[i] Removing the build directory.")
-shutil.rmtree('lingpy_build/')
-print("[i] Done.")
-print("[i] LingPy was successfully installed on your system.")
+if 'install' in sys.argv:
+    print("[i] Removing the build directory.")
+    shutil.rmtree('lingpy_build/')
+    print("[i] Done.")
+    print("[i] LingPy was successfully installed on your system.")
